pyproject.toml
[project]
name = "ai-agents-and-workflows"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "openai>=2.6.1",
    "python-dotenv>=1.2.1",
]

[tool.pyright]
venvPath = "."
venv = "venv"
pythonVersion = "3.13"
pythonPlatform = "Darwin"
